+++ mgnipy/V2/query_set.py
@@ -14,13 +14,19 @@ import pandas as pd
 
 from mgnipy._models.config import MgnipyConfig
 from mgnipy._models.CONSTANTS import SupportedEndpoints
+from mgnipy.V2.endpoints import (
+    ACC_DETAIL_ENDPOINTS,
+)
 from mgnipy.V2.endpoints import ALL_ENDPOINTS as ENDPOINTS
+from mgnipy.V2.endpoints import (
+    BETWEEN_RESOURCE_RELATIONSHIPS,
+    WITHIN_RESOURCE_RELATIONSHIPS,
+)
 from mgnipy.V2.mixins import (
     ResultHandlerMixin,
 )
 from mgnipy.V2.query_executor import QueryExecutor
 
-BASE_URL = MgnipyConfig().base_url
 ID_PARAM = {
     SupportedEndpoints.BIOMES: "biome_lineage",
     SupportedEndpoints.BIOME: "biome_lineage",
@@ -69,11 +75,14 @@ class QuerySet(ResultHandlerMixin):
             "assembly",
         ],
         *,
+        config: Optional[MgnipyConfig] = None,
         params: Optional[dict[str, Any]] = None,
         **kwargs,
     ):
 
-        self._base_url: str = BASE_URL
+        self.config = MgnipyConfig(**config) if config else MgnipyConfig()
+
+        self._base_url: str = str(self.config.base_url)
         self._resource = SupportedEndpoints.validate(resource)
         # params as dict
         self._params: dict[str, Any] = params or {}
@@ -533,3 +542,97 @@ class QuerySet(ResultHandlerMixin):
             raise AttributeError(
                 f"Identifier key '{self.id_param_key}' not found in parameters for resource '{self.resource}'."
             ) from None
+
+    # RELATIONSHIP HANDLING
+    def list_relationships(self) -> list[str]:
+        if self.resource in WITHIN_RESOURCE_RELATIONSHIPS:
+            return [WITHIN_RESOURCE_RELATIONSHIPS[self.resource].value]
+        elif self.resource in BETWEEN_RESOURCE_RELATIONSHIPS:
+            return [
+                endpoint.value
+                for endpoint in BETWEEN_RESOURCE_RELATIONSHIPS[self.resource]
+            ]
+        else:
+            return []
+
+    def _get_next_rel_resource(self, name: Optional[str] = None) -> SupportedEndpoints:
+        """
+        Get the next resource name based on the relationship name
+        or just list the first or only endpoint.
+        """
+        if name is not None and name in self.list_relationships():
+            return SupportedEndpoints.validate(name)
+        if name is not None and name not in self.list_relationships():
+            raise AttributeError(
+                f"{self.resource} does not have a linked resource {name}."
+            )
+        if name is None and len(self.list_relationships()) > 0:
+            return SupportedEndpoints.validate(self.list_relationships()[0])
+        raise AttributeError(f"{self.resource} does not have any linked resources.")
+
+    def next_rel_module(self, name: Optional[str] = None) -> Callable:
+        # get SupportedEndpoint of next resource
+        next_resource = self._get_next_rel_resource(name)
+        # then get the endpoint function for that resource
+        if self.resource in WITHIN_RESOURCE_RELATIONSHIPS:
+            return ACC_DETAIL_ENDPOINTS[next_resource]
+        elif self.resource in BETWEEN_RESOURCE_RELATIONSHIPS:
+            return BETWEEN_RESOURCE_RELATIONSHIPS[self.resource][next_resource]
+        else:
+            raise AttributeError(
+                f"{self.resource.value} does not have any linked resources."
+            )
+
+    def _resolve_access_param(self, key: int | str) -> dict:
+        # allow index-based access
+        if self.results_ids is not None and isinstance(key, int):
+            return {self.id_param_key: self.results_ids[key]}
+        # or by accession/biome_lineage/ids string directly
+        if self.results_ids is not None and key in self.results_ids:
+            return {self.id_param_key: key}
+
+        raise KeyError(
+            f"Invalid key: {key}. "
+            "Key must be an integer index, or a valid id string. "
+            "Accession/id/biome_lineage must exist in`.results_ids`"
+        )
+
+    def get_next(
+        self,
+        access_param: dict[str, str],
+        resource_name: Optional[str] = None,
+        fetch: bool = True,
+    ) -> "QuerySet":
+        """
+        Independent of list vs detail resource, get next linked proxy for a specific accession.
+
+            Parameters
+            ----------
+            access_param : dict[str, str]
+                A dictionary containing the necessary parameter to identify the detail resource,
+                such as {"accession": "MGYS00001234"} or {"biome_lineage": "root"}.
+            resource_name : Optional[str]
+                The name of the resource to get the next instance of. If None, will use the first or only linked resource.
+            fetch : bool
+                Whether to immediately fetch the detail after creating the proxy.
+
+
+            Returns
+            -------
+            QuerySet
+                A proxy for the next resource.
+
+            Examples
+            -------
+            sample = samples.getting_next({"accession": "MGYS00001234"})
+            samples = study.getting_next({"accession": "MGYS00001234"})
+        """
+
+        next_resource = self.next_resource(resource_name).value
+        next_module = self.next_resource_module(resource_name)
+
+        child = self._clone(resource=next_resource, **access_param)
+        child.endpoint_module = next_module
+        if fetch:
+            child.get(safety=False)
+        return child
